+++ source/Public/Clear-DatumRsopCache.ps1
@@ -1,0 +1,12 @@
+function Clear-DatumRsopCache
+{
+    [CmdletBinding()]
+
+    param ()
+
+    if ($rsopCache.Count)
+    {
+        $rsopCache.Clear()
+        Write-Verbose -Message 'Datum RSOP Cache cleared'
+    }
+}